ora conn pool is managed, add request queue
when oraSock is required but not available, the request can be queued, when there are new free oraSock available, the request will be called with that oraSock.

Noradle will always choose the most used and recently used oraSock from pool, so oracle will execute the work using constantly executing process, that will help to avoid process switch.